NTBS-2284 Refactoring
