+++ load-tests/README.md
@@ -3,5 +3,6 @@
 ## Initial setup
 
 In order to run the NTBS load tests, you must do the following:
+
 1. Download Gatling Open Source from [https://gatling.io/open-source/start-testing/](https://gatling.io/open-source/start-testing/).
 1. Unzip the bundle and copy the `bin`, `lib` and `results` directories into the root directory of the load testing project.
