+++ load-tests/user-files/simulations/ScenarioBuilders/DashboardScenarioBuilder.scala
@@ -1,0 +1,10 @@
+import io.gatling.core.Predef._
+import io.gatling.http.Predef._
+import io.gatling.jdbc.Predef._
+import io.gatling.core.structure.{ StructureBuilder, ChainBuilder }
+
+object DashboardScenarioBuilder {
+    def build(): StructureBuilder[ChainBuilder] = {
+        exec(NtbsRequestBuilder.getRequest("dashboard", "/")).pause(2)
+    }
+}
